pyproject.toml
[tool.poetry]
authors = ["Patrick-Oliver Groß <pa.gross@fme.de>"]
description = "A cli tool to simplify the use of the Qlik Engine JSON API"
name = "qlik-engine-helper"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.8"
python-dateutil = "^2.8.1"
tabulate = "^0.8.7"
websockets = "^8.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.3"
isort = "^5.4.2"
mypy = "^0.782"
pre-commit = "^2.9.2"
pytest = "^5.2"

[tool.poetry.scripts]
qlik-engine-helper = "qlik_engine_helper.main:main"

[tool.black]
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )
)
'''
include = '\.pyi?$'
line-length = 88
target-version = ['py37', 'py38']

[tool.isort]
known_third_party = "websockets, typer"
multi_line_output = 3
profile = "black"

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
